+++ ErsatzTV.Application/FFmpegProfiles/Commands/UpdateFFmpegSettingsHandler.cs
@@ -83,7 +83,7 @@ public class UpdateFFmpegSettingsHandler : IRequestHandler<UpdateFFmpegSettings,
 
         await _configElementRepository.Upsert(
             ConfigElementKey.FFmpegPreferredLanguageCode,
-            request.Settings.PreferredLanguageCode);
+            request.Settings.PreferredAudioLanguageCode);
 
         if (request.Settings.GlobalWatermarkId is not null)
         {