+++ ErsatzTV.FFmpeg/Filter/ComplexFilter.cs
@@ -9,6 +9,7 @@ public class ComplexFilter : IPipelineStep
     private readonly Option<VideoInputFile> _maybeVideoInputFile;
     private readonly Option<AudioInputFile> _maybeAudioInputFile;
     private readonly Option<WatermarkInputFile> _maybeWatermarkInputFile;
+    private readonly Option<SubtitleInputFile> _maybeSubtitleInputFile;
     private readonly FrameSize _resolution;
 
     public ComplexFilter(
@@ -17,6 +18,7 @@ public class ComplexFilter : IPipelineStep
         Option<VideoInputFile> maybeVideoInputFile,
         Option<AudioInputFile> maybeAudioInputFile,
         Option<WatermarkInputFile> maybeWatermarkInputFile,
+        Option<SubtitleInputFile> maybeSubtitleInputFile,
         FrameSize resolution)
     {
         _currentState = currentState;
@@ -24,6 +26,7 @@ public class ComplexFilter : IPipelineStep
         _maybeVideoInputFile = maybeVideoInputFile;
         _maybeAudioInputFile = maybeAudioInputFile;
         _maybeWatermarkInputFile = maybeWatermarkInputFile;
+        _maybeSubtitleInputFile = maybeSubtitleInputFile;
         _resolution = resolution;
     }
 
@@ -32,13 +35,16 @@ public class ComplexFilter : IPipelineStep
         var audioLabel = "0:a";
         var videoLabel = "0:v";
         string watermarkLabel;
+        string subtitleLabel;
 
         var result = new List<string>();
 
         string audioFilterComplex = string.Empty;
         string videoFilterComplex = string.Empty;
         string watermarkFilterComplex = string.Empty;
-        string overlayFilterComplex = string.Empty;
+        string watermarkOverlayFilterComplex = string.Empty;
+        string subtitleFilterComplex = string.Empty;
+        string subtitleOverlayFilterComplex = string.Empty;
         
         var distinctPaths = new List<string>();
         foreach ((string path, _) in _maybeVideoInputFile)
@@ -58,6 +64,14 @@ public class ComplexFilter : IPipelineStep
         }
 
         foreach ((string path, _) in _maybeWatermarkInputFile)
+        {
+            if (!distinctPaths.Contains(path))
+            {
+                distinctPaths.Add(path);
+            }
+        }
+
+        foreach ((string path, _) in _maybeSubtitleInputFile)
         {
             if (!distinctPaths.Contains(path))
             {
@@ -117,7 +131,7 @@ public class ComplexFilter : IPipelineStep
                     watermarkFilterComplex += watermarkLabel;
                 }
 
-                IPipelineFilterStep overlayFilter = AvailableOverlayFilters.ForAcceleration(
+                IPipelineFilterStep overlayFilter = AvailableWatermarkOverlayFilters.ForAcceleration(
                     _ffmpegState.HardwareAccelerationMode,
                     _currentState,
                     watermarkInputFile.DesiredState,
@@ -141,10 +155,57 @@ public class ComplexFilter : IPipelineStep
                         uploadFilter = "," + uploadFilter;
                     }
                     
-                    overlayFilterComplex = $"{tempVideoLabel}{watermarkLabel}{overlayFilter.Filter}{uploadFilter}[vf]";
+                    watermarkOverlayFilterComplex = $"{tempVideoLabel}{watermarkLabel}{overlayFilter.Filter}{uploadFilter}[vf]";
                     
                     // change the mapped label
                     videoLabel = "[vf]";
+                }
+            }
+        }
+        
+        foreach (SubtitleInputFile subtitleInputFile in _maybeSubtitleInputFile.Filter(s => !s.Copy))
+        {
+            int inputIndex = distinctPaths.IndexOf(subtitleInputFile.Path);
+            foreach ((int index, _, _) in subtitleInputFile.Streams)
+            {
+                subtitleLabel = $"{inputIndex}:{index}";
+                if (subtitleInputFile.FilterSteps.Any(f => !string.IsNullOrWhiteSpace(f.Filter)))
+                {
+                    subtitleFilterComplex += $"[{inputIndex}:{index}]";
+                    subtitleFilterComplex += string.Join(
+                        ",",
+                        subtitleInputFile.FilterSteps.Select(f => f.Filter).Filter(s => !string.IsNullOrWhiteSpace(s)));
+                    subtitleLabel = "[st]";
+                    subtitleFilterComplex += subtitleLabel;
+                }
+
+                IPipelineFilterStep overlayFilter =
+                    AvailableSubtitleOverlayFilters.ForAcceleration(_ffmpegState.HardwareAccelerationMode);
+
+                if (overlayFilter.Filter != string.Empty)
+                {
+                    string tempVideoLabel = string.IsNullOrWhiteSpace(videoFilterComplex) &&
+                                            string.IsNullOrWhiteSpace(watermarkFilterComplex)
+                        ? $"[{videoLabel}]"
+                        : videoLabel;
+
+                    // vaapi uses software overlay and needs to upload
+                    string uploadFilter = string.Empty;
+                    if (_ffmpegState.HardwareAccelerationMode == HardwareAccelerationMode.Vaapi)
+                    {
+                        uploadFilter = new HardwareUploadFilter(_ffmpegState).Filter;
+                    }
+
+                    if (!string.IsNullOrWhiteSpace(uploadFilter))
+                    {
+                        uploadFilter = "," + uploadFilter;
+                    }
+
+                    subtitleOverlayFilterComplex =
+                        $"{tempVideoLabel}{subtitleLabel}{overlayFilter.Filter}{uploadFilter}[vst]";
+
+                    // change the mapped label
+                    videoLabel = "[vst]";
                 }
             }
         }
@@ -153,13 +214,31 @@ public class ComplexFilter : IPipelineStep
         {
             var filterComplex = string.Join(
                 ";",
-                new[] { audioFilterComplex, videoFilterComplex, watermarkFilterComplex, overlayFilterComplex }.Where(
+                new[]
+                {
+                    audioFilterComplex,
+                    videoFilterComplex,
+                    watermarkFilterComplex,
+                    subtitleFilterComplex,
+                    watermarkOverlayFilterComplex,
+                    subtitleOverlayFilterComplex
+                }.Where(
                     s => !string.IsNullOrWhiteSpace(s)));
 
             result.AddRange(new[] { "-filter_complex", filterComplex });
         }
 
         result.AddRange(new[] { "-map", audioLabel, "-map", videoLabel });
+
+        foreach (SubtitleInputFile subtitleInputFile in _maybeSubtitleInputFile.Filter(s => s.Copy))
+        {
+            int inputIndex = distinctPaths.IndexOf(subtitleInputFile.Path);
+            foreach ((int index, _, _) in subtitleInputFile.Streams)
+            {
+                subtitleLabel = $"{inputIndex}:{index}";
+                result.AddRange(new[] { "-map", subtitleLabel });
+            }
+        }
 
         return result;
     }