+++ ErsatzTV.Infrastructure/Migrations/20220329005536_Add_FFmpegProfileSubtitleMode.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using ErsatzTV.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
@@ -10,9 +11,10 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace ErsatzTV.Infrastructure.Migrations
 {
     [DbContext(typeof(TvContext))]
-    partial class TvContextModelSnapshot : ModelSnapshot
+    [Migration("20220329005536_Add_FFmpegProfileSubtitleMode")]
+    partial class Add_FFmpegProfileSubtitleMode
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder.HasAnnotation("ProductVersion", "6.0.3");
@@ -239,16 +241,10 @@ namespace ErsatzTV.Infrastructure.Migrations
                     b.Property<string>("Number")
                         .HasColumnType("TEXT");
 
-                    b.Property<string>("PreferredAudioLanguageCode")
-                        .HasColumnType("TEXT");
-
-                    b.Property<string>("PreferredSubtitleLanguageCode")
+                    b.Property<string>("PreferredLanguageCode")
                         .HasColumnType("TEXT");
 
                     b.Property<int>("StreamingMode")
-                        .HasColumnType("INTEGER");
-
-                    b.Property<int>("SubtitleMode")
                         .HasColumnType("INTEGER");
 
                     b.Property<Guid>("UniqueId")
@@ -531,6 +527,9 @@ namespace ErsatzTV.Infrastructure.Migrations
                         .HasColumnType("INTEGER");
 
                     b.Property<int>("ResolutionId")
+                        .HasColumnType("INTEGER");
+
+                    b.Property<int>("SubtitleMode")
                         .HasColumnType("INTEGER");
 
                     b.Property<int>("ThreadCount")