+++ ErsatzTV/ViewModels/ChannelEditViewModel.cs
@@ -11,11 +11,13 @@ public class ChannelEditViewModel
     public string Categories { get; set; }
     public string Number { get; set; }
     public int FFmpegProfileId { get; set; }
-    public string PreferredLanguageCode { get; set; }
+    public string PreferredAudioLanguageCode { get; set; }
     public string Logo { get; set; }
     public StreamingMode StreamingMode { get; set; }
     public int? WatermarkId { get; set; }
     public int? FallbackFillerId { get; set; }
+    public string PreferredSubtitleLanguageCode { get; set; }
+    public ChannelSubtitleMode SubtitleMode { get; set; }
 
     public UpdateChannel ToUpdate() =>
         new(
@@ -26,10 +28,12 @@ public class ChannelEditViewModel
             Categories,
             FFmpegProfileId,
             Logo,
-            PreferredLanguageCode,
+            PreferredAudioLanguageCode,
             StreamingMode,
             WatermarkId,
-            FallbackFillerId);
+            FallbackFillerId,
+            PreferredSubtitleLanguageCode,
+            SubtitleMode);
 
     public CreateChannel ToCreate() =>
         new(
@@ -39,8 +43,10 @@ public class ChannelEditViewModel
             Categories,
             FFmpegProfileId,
             Logo,
-            PreferredLanguageCode,
+            PreferredAudioLanguageCode,
             StreamingMode,
             WatermarkId,
-            FallbackFillerId);
+            FallbackFillerId,
+            PreferredSubtitleLanguageCode,
+            SubtitleMode);
 }